created Foo, ServiceRegistry, Service and Driver classes
Foo is the implementation of CircuitBreaker which can handle multiple services and has routeRequest method which sends requests to specific services based on their current state(CLOSED, OPEN, PARTIALLY_OPEN). | ServiceRegistry holds the all the services for one instance of CircuitBreaker. It does periodic health checks on registered services and based on the failed and successful health checks, sets the state of a service. | Service contains the url to the health endpoint of that service and sends a get request to it. If responseCode to health endpoint is 200 Ok, then only successfulRequestCount is incremented otherwise the failedRequest count is incremented for that service. It also contains the handleRequest method with makes the actual request to the service.